+++ R/data.R
@@ -1,22 +1,29 @@
-#' Adverse Events Dataset
+#' Adverse Events Dataset-updated
+#'
+#' An updated SDTM AE dataset that uses the CDISC pilot project
+#'
+#' @source \url{https://github.com/pharmaverse/admiral.test/blob/main/data/admiral_ae.rda}
+"admiral_ae"
+
+#' Adverse Events Dataset-raw
 #'
 #' A SDTM AE dataset from the CDISC pilot project
 #'
-#' @source \url{https://github.com/cdisc-org/sdtm-adam-pilot-project/blob/master/updated-pilot-submission-package/900172/m5/datasets/cdiscpilot01/tabulations/sdtm/ae.xpt?raw=true}
-"admiral_ae"
+#' @source \url{https://github.com/cdisc-org/sdtm-adam-pilot-project/blob/master/updated-pilot-submission-package/900172/m5/datasets/cdiscpilot01/tabulations/sdtm/ae.xpt?raw=true} # nolint
+"raw_ae"
 
 #' Concomitant Medication Dataset
 #'
 #' A SDTM CM dataset from the CDISC pilot project
 #'
-#' @source \url{https://github.com/cdisc-org/sdtm-adam-pilot-project/blob/master/updated-pilot-submission-package/900172/m5/datasets/cdiscpilot01/tabulations/sdtm/cm.xpt?raw=true}
+#' @source \url{https://github.com/cdisc-org/sdtm-adam-pilot-project/blob/master/updated-pilot-submission-package/900172/m5/datasets/cdiscpilot01/tabulations/sdtm/cm.xpt?raw=true} # nolint
 "admiral_cm"
 
 #' Demography Dataset
 #'
 #' A SDTM DM dataset from the CDISC pilot project
 #'
-#' @source \url{https://github.com/cdisc-org/sdtm-adam-pilot-project/blob/master/updated-pilot-submission-package/900172/m5/datasets/cdiscpilot01/tabulations/sdtm/dm.xpt?raw=true}
+#' @source \url{https://github.com/cdisc-org/sdtm-adam-pilot-project/blob/master/updated-pilot-submission-package/900172/m5/datasets/cdiscpilot01/tabulations/sdtm/dm.xpt?raw=true} # nolint
 "admiral_dm"
 
 #' Disposition Dataset-updated
@@ -31,7 +38,7 @@
 #'
 #' A SDTM DS dataset from the CDISC pilot project
 #'
-#' @source \url{https://github.com/cdisc-org/sdtm-adam-pilot-project/blob/master/updated-pilot-submission-package/900172/m5/datasets/cdiscpilot01/tabulations/sdtm/ds.xpt?raw=true}
+#' @source \url{https://github.com/cdisc-org/sdtm-adam-pilot-project/blob/master/updated-pilot-submission-package/900172/m5/datasets/cdiscpilot01/tabulations/sdtm/ds.xpt?raw=true} # nolint
 "raw_ds"
 
 #' Electrocardiogram Dataset
@@ -44,7 +51,7 @@
 #'
 #' A SDTM EX dataset from the CDISC pilot project
 #'
-#' @source \url{https://github.com/cdisc-org/sdtm-adam-pilot-project/blob/master/updated-pilot-submission-package/900172/m5/datasets/cdiscpilot01/tabulations/sdtm/ex.xpt?raw=true}
+#' @source \url{https://github.com/cdisc-org/sdtm-adam-pilot-project/blob/master/updated-pilot-submission-package/900172/m5/datasets/cdiscpilot01/tabulations/sdtm/ex.xpt?raw=true} # nolint
 "admiral_ex"
 
 #' Laboratory Measurements Dataset
@@ -59,14 +66,14 @@
 #'
 #' A SDTM LB dataset from the CDISC pilot project
 #'
-#' @source \url{https://github.com/cdisc-org/sdtm-adam-pilot-project/blob/master/updated-pilot-submission-package/900172/m5/datasets/cdiscpilot01/tabulations/sdtm/lb.xpt?raw=true}
+#' @source \url{https://github.com/cdisc-org/sdtm-adam-pilot-project/blob/master/updated-pilot-submission-package/900172/m5/datasets/cdiscpilot01/tabulations/sdtm/lb.xpt?raw=true} # nolint
 "raw_lb"
 
 #' Medical History Dataset-updated
 #'
 #' An updated SDTM MH dataset that uses data from the CDISC pilot project
 #'
-#' @source \url{https://github.com/cdisc-org/sdtm-adam-pilot-project/blob/master/updated-pilot-submission-package/900172/m5/datasets/cdiscpilot01/tabulations/sdtm/mh.xpt?raw=true}
+#' @source \url{https://github.com/cdisc-org/sdtm-adam-pilot-project/blob/master/updated-pilot-submission-package/900172/m5/datasets/cdiscpilot01/tabulations/sdtm/mh.xpt?raw=true} # nolint
 #' @author Annie Yang
 "admiral_mh"
 
@@ -74,7 +81,7 @@
 #'
 #' A SDTM MH dataset from the CDISC pilot project
 #'
-#' @source \url{https://github.com/cdisc-org/sdtm-adam-pilot-project/blob/master/updated-pilot-submission-package/900172/m5/datasets/cdiscpilot01/tabulations/sdtm/mh.xpt?raw=true}
+#' @source \url{https://github.com/cdisc-org/sdtm-adam-pilot-project/blob/master/updated-pilot-submission-package/900172/m5/datasets/cdiscpilot01/tabulations/sdtm/mh.xpt?raw=true} # nolint
 "raw_mh"
 
 #' Pharmacokinetics Concentrations Dataset
@@ -97,7 +104,7 @@
 #'
 #' A SDTM QS dataset from the CDISC pilot project
 #'
-#' @source \url{https://github.com/cdisc-org/sdtm-adam-pilot-project/blob/master/updated-pilot-submission-package/900172/m5/datasets/cdiscpilot01/tabulations/sdtm/qs.xpt?raw=true}
+#' @source \url{https://github.com/cdisc-org/sdtm-adam-pilot-project/blob/master/updated-pilot-submission-package/900172/m5/datasets/cdiscpilot01/tabulations/sdtm/qs.xpt?raw=true} # nolint
 "admiral_qs"
 
 #' Tumor Identification Dataset
@@ -125,28 +132,28 @@
 #'
 #' A SDTM SUPPAE dataset from the CDISC pilot project
 #'
-#' @source \url{https://github.com/cdisc-org/sdtm-adam-pilot-project/blob/master/updated-pilot-submission-package/900172/m5/datasets/cdiscpilot01/tabulations/sdtm/suppae.xpt?raw=true}
+#' @source \url{https://github.com/cdisc-org/sdtm-adam-pilot-project/blob/master/updated-pilot-submission-package/900172/m5/datasets/cdiscpilot01/tabulations/sdtm/suppae.xpt?raw=true} # nolint
 "admiral_suppae"
 
 #' Supplemental Disposition Dataset-updated
 #'
 #' A SDTM SUPPDS dataset from the CDISC pilot project
 #'
-#' @source \url{https://github.com/cdisc-org/sdtm-adam-pilot-project/blob/master/updated-pilot-submission-package/900172/m5/datasets/cdiscpilot01/tabulations/sdtm/suppds.xpt?raw=true}
+#' @source \url{https://github.com/cdisc-org/sdtm-adam-pilot-project/blob/master/updated-pilot-submission-package/900172/m5/datasets/cdiscpilot01/tabulations/sdtm/suppds.xpt?raw=true} # nolint
 "admiral_suppds"
 
 #' Supplemental Disposition Dataset-raw
 #'
 #' A SDTM SUPPDS dataset from the CDISC pilot project
 #'
-#' @source \url{https://github.com/cdisc-org/sdtm-adam-pilot-project/blob/master/updated-pilot-submission-package/900172/m5/datasets/cdiscpilot01/tabulations/sdtm/suppds.xpt?raw=true}
+#' @source \url{https://github.com/cdisc-org/sdtm-adam-pilot-project/blob/master/updated-pilot-submission-package/900172/m5/datasets/cdiscpilot01/tabulations/sdtm/suppds.xpt?raw=true} # nolint
 "raw_suppds"
 
 #' Supplemental Demography Dataset
 #'
 #' A SDTM SUPPDM dataset from the CDISC pilot project
 #'
-#' @source \url{https://github.com/cdisc-org/sdtm-adam-pilot-project/blob/master/updated-pilot-submission-package/900172/m5/datasets/cdiscpilot01/tabulations/sdtm/suppdm.xpt?raw=true}
+#' @source \url{https://github.com/cdisc-org/sdtm-adam-pilot-project/blob/master/updated-pilot-submission-package/900172/m5/datasets/cdiscpilot01/tabulations/sdtm/suppdm.xpt?raw=true} # nolint
 "admiral_suppdm"
 
 #' Supplemental Tumor Results Dataset
@@ -160,14 +167,14 @@
 #'
 #' A SDTM TS dataset from the CDISC pilot project
 #'
-#' @source \url{https://github.com/cdisc-org/sdtm-adam-pilot-project/blob/master/updated-pilot-submission-package/900172/m5/datasets/cdiscpilot01/tabulations/sdtm/ts.xpt?raw=true}
+#' @source \url{https://github.com/cdisc-org/sdtm-adam-pilot-project/blob/master/updated-pilot-submission-package/900172/m5/datasets/cdiscpilot01/tabulations/sdtm/ts.xpt?raw=true} # nolint
 "admiral_ts"
 
 #' Vital Signs Dataset
 #'
 #' A SDTM VS dataset from the CDISC pilot project
 #'
-#' @source \url{https://github.com/cdisc-org/sdtm-adam-pilot-project/blob/master/updated-pilot-submission-package/900172/m5/datasets/cdiscpilot01/tabulations/sdtm/vs.xpt?raw=true}
+#' @source \url{https://github.com/cdisc-org/sdtm-adam-pilot-project/blob/master/updated-pilot-submission-package/900172/m5/datasets/cdiscpilot01/tabulations/sdtm/vs.xpt?raw=true} # nolint
 "admiral_vs"
 
 #' SMQ Dataset
@@ -186,5 +193,5 @@
 #'
 #' A SDTM SV dataset from the CDISC pilot project
 #'
-#' @source \url{https://github.com/cdisc-org/sdtm-adam-pilot-project/blob/master/updated-pilot-submission-package/900172/m5/datasets/cdiscpilot01/tabulations/sdtm/sv.xpt?raw=true}
+#' @source \url{https://github.com/cdisc-org/sdtm-adam-pilot-project/blob/master/updated-pilot-submission-package/900172/m5/datasets/cdiscpilot01/tabulations/sdtm/sv.xpt?raw=true} # nolint
 "admiral_sv"