+++ NEWS.md
@@ -1,3 +1,10 @@
+# admiral.test 0.5.0
+ - Updated AE to add variable AELAT for ophthalmology package (#99)
+
+# admiral.test 0.3.0
+ - Updated labels for pp and pc datasets (#77)
+ - Implemented admiral CI/CD workflows
+
 # admiral.test 0.2.0
 
 - Renaming all datasets to `admiral_` (#30)